+++ massiv-examples/vision/Setup.hs
@@ -3,4 +3,3 @@ import Distribution.Simple
 main :: IO ()
 main = defaultMain
 #endif
-